% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean Earthquake Data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{A data frame of NOAA significant earthquake data}
}
\value{
A \code{tbl_df} with the same supplied data, but with
  \code{LATITUDE}, \code{LONGITUDE}, \code{DEATHS}, \code{TOTAL_DEATHS}, and
  \code{EQ_PRIMARY} converted from \code{character} to \code{numeric}, and a
  new column \code{DATE}.
}
\description{
\code{eq_clean_data} takes the raw data frame of NOAA earthquake and cleans it
}
\details{
This function takes the raw data frame of NOAA earthquake data and cleans it.
Data (tab delimited) is available for download at:
\href{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}{Significant
EarthQuake Database}
The clean data frame have the following:
1. A date column created by uniting the year, month, day and converting it to
   the Date class
2. LATITUDE and LONGITUDE columns converted to numeric class
}
\examples{

data(earthquakes)
quakes_clean <- eq_clean_data(earthquakes)
}
\author{
Federico Sensi
}
