% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline - adds an earthquake timeline plot to a ggplot2/grid graphics objects}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or \code{\link{aes_}}.
If specified and \code{inherit.aes = TRUE} (the default), it is combined with the default mapping at the top level of the plot.
Required aesthetics are: \code{x} (usually center longitude), \code{y} (usually center latitude), \code{r_ne}, \code{r_se}, \code{r_sw}, \code{r_nw}  (the northeast, southeast, southwest, and northwest wind radius measurements, respectively), \code{fill}, \code{color} (values to use for fill and color, usually wind speed for the measurements)
Optional aesthetics are: \code{scale_radii} (optional scaling factor for the wind radii, default is \code{1.0})}

\item{data}{The data to be displayed in this layer. There are three options:
If \code{NULL}, the default, the data is inherited from the plot data as specified in the call to \code{\link{ggplot}}.
A \code{data.frame}, or other object, will override the plot data. All objects will be fortified to produce a data frame.
See \code{\link{fortify}} for which variables will be created.
A \code{function} will be called with a single argument, the plot data. The return value must be a \code{data.frame.},
and will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with them.}
}
\value{
This function returns a \code{ggplot2::layer}.
}
\description{
\code{geom_timeline} shows a timeline of NOAA Significant earthquakes,
plotting individual countries along the y-axis and dates along the x-axis.
The size of the points is relatative to the earthquakes' magnitude, and
the color is related to the total number of deaths.
}
\examples{

 library(magrittr)
 library(ggplot2)
 library(grid)

data(earthquakes)
quakes_clean <- eq_clean_data(earthquakes)

 quakes_clean \%>\%
    dplyr::filter(COUNTRY \%in\% c('ITALY', 'GREECE') & DATE > '1990-01-01' & !is.na(EQ_PRIMARY)) \%>\%
    ggplot() +
    geom_timeline(aes(x = DATE, y = COUNTRY, color = TOTAL_DEATHS,
                      size = EQ_PRIMARY)) +
    scale_size_continuous(name = 'Richter scale value') +
    scale_color_continuous(name = 'Nbr of Deaths')
}
\author{
Federico Sensi
}
