% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID3.R
\docType{data}
\name{ID3}
\alias{ID3}
\title{ID3}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ID3
}
\description{
A Decision Tree implemented using the ID3 algorithm.
}
\examples{
clf <- ID3$new("monks")
clf$train(monks1_train[, -c(1,8)], monks1_train[,1])
clf$predict(monks1_test[, -c(1,8)])
print(clf$tree, "feature")
plot(clf$tree)

}
\keyword{datasets}
