% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monks2_train}
\alias{monks2_train}
\title{Monk's Problems Data Set (Problem 2 Training Set)}
\format{A data frame with 169 rows and 8 columns:
\describe{
  \item{label}{The class label}
  \item{a1}{Attribute 1}
  \item{a2}{Attribute 2}
  \item{a3}{Attribute 3}
  \item{a4}{Attribute 4}
  \item{a5}{Attribute 5}
  \item{a6}{Attribute 6}
  \item{id}{A unique identification number}
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/MONK's+Problems}
}
\usage{
monks2_train
}
\description{
The training set for problem 2 of the Monk's Problems Data Set,
available from the UCI Machine Learning Repository.
}
\keyword{datasets}
